fix(eval): scalar results for rank-0 ∊/⊇, long results for monadic × on doubles
- membership: rank-0 (enclosed) left arg returns scalar, result keeps
  left shape (oracle: (⊂1 2) ∊ 10 11 12 → 0, ⍴ → ⍬)
- pick_apl: scalar/rank-0 index returns disclosed element
  (oracle: 2 ⊇ 10 20 30 → 30, 1 + (0 ⊇ 2 3 4 5 6 7) → 3)
- KapNumber::signum: Double arm returns Long (sole caller is monadic ×;
  oracle: × 1.1 ¯1.6 0.0 → ⟨1 -1 0⟩, × 0.5 → 1)
- 3 stale curated ⊇ expectations fixed, 6 new curated regressions

+++ kap-core/src/number.rs
@@ -1026,7 +1026,7 @@ impl KapNumber {
         use KapNumber::*;
         match self {
             Long(v) => Long(if *v > 0 { 1 } else if *v < 0 { -1 } else { 0 }),
-            Double(v) => Double(if *v > 0.0 { 1.0 } else if *v < 0.0 { -1.0 } else { 0.0 }),
+            Double(v) => Long(if *v > 0.0 { 1 } else if *v < 0.0 { -1 } else { 0 }),
             BigInt(v) => Long(if *v > num_bigint::BigInt::from(0) { 1 } else if *v < num_bigint::BigInt::from(0) { -1 } else { 0 }),
             Rational(v) => Long(if *v.numer() > num_bigint::BigInt::from(0) { 1 } else if *v.numer() < num_bigint::BigInt::from(0) { -1 } else { 0 }),
             Complex(r, i) => {
